Simple Cafe Billing System.
This program can compute the total bill of a customer digitally.
This program is built using only Python Language.

tkinter framework

This system contains a simple GUI application that uses tkinter framework. The user can only enter the quantity of an items ordered by the customers. The system can display the list of item price when clicked price button, and also when the item is ordered it generates the total bill of the purchase
item by clicking Total button. The Order no is auto generated after purchase.